+++ Tests/Unit/SharePointDsc/SharePointDsc.SPFarm.Tests.ps1
@@ -120,14 +120,6 @@ namespace Microsoft.SharePoint.Administration {
                 RunCentralAdmin           = $true
             }
 
-            It "Should throw exception in the get method" {
-                { Get-TargetResource @testParams } | Should Throw "CentralAdministrationUrl is not a valid URI. It should include the scheme (http/https) and address."
-            }
-
-            It "Should throw exception in the test method" {
-                { Test-TargetResource @testParams } | Should Throw "CentralAdministrationUrl is not a valid URI. It should include the scheme (http/https) and address."
-            }
-
             It "Should throw exception in the set method" {
                 { Set-TargetResource @testParams } | Should Throw "CentralAdministrationUrl is not a valid URI. It should include the scheme (http/https) and address."
             }
@@ -147,14 +139,6 @@ namespace Microsoft.SharePoint.Administration {
                 RunCentralAdmin           = $true
             }
 
-            It "Should throw exception in the get method" {
-                { Get-TargetResource @testParams } | Should Throw "CentralAdministrationUrl should not specify port. Use CentralAdministrationPort instead."
-            }
-
-            It "Should throw exception in the test method" {
-                { Test-TargetResource @testParams } | Should Throw "CentralAdministrationUrl should not specify port. Use CentralAdministrationPort instead."
-            }
-
             It "Should throw exception in the set method" {
                 { Set-TargetResource @testParams } | Should Throw "CentralAdministrationUrl should not specify port. Use CentralAdministrationPort instead."
             }
@@ -172,14 +156,6 @@ namespace Microsoft.SharePoint.Administration {
                 Passphrase                = $mockPassphrase
                 AdminContentDatabaseName  = "SP_AdminContent"
                 RunCentralAdmin           = $true
-            }
-
-            It "Should throw exception in the get method" {
-                { Get-TargetResource @testParams } | Should Throw "CentralAdministrationUrl parameter can only be used with HTTPS"
-            }
-
-            It "Should throw exception in the test method" {
-                { Test-TargetResource @testParams } | Should Throw "CentralAdministrationUrl parameter can only be used with HTTPS"
             }
 
             It "Should throw exception in the set method" {